Split xml_parser by tasks
